Prices for life insurance

The AARP (original named the "American Association of Retired People" but now just AARP) offers life insurance to it's members. The data come from a full-page advertisement (circa 2012) in the "AARP Bulletin", which has the second largest circulation in the world of any magazine, with upward of 40 million subscribers. (Only the "AARP Magazine" has a larger circulation.)

data(AARP)

Format

A data frame with 36 observations on the following variables.

Age The age of the person covered by the insurance policy.

Sex The sex of the person covered by the insurance policy.

Coverage The "death benefit" in 1000 USD.

Cost Monthly cost in USD.

Source

The "AARP Bulletin". A copy of the ad is available at this link.

Details

Life insurance provides a "death benefit", money paid out to the insured person's survivors upon death of the insured. There is a cost for the insurance. Among other factors, the cost depends on both age and sex. (For this type of insurance, called "term insurance", the cost changes as the insured person ages.)

Examples

mod_1 <- lm(Cost ~ Age + Coverage, data = AARP)
mod_effect(mod_1, ~ Coverage)

#>      slope Coverage to_Coverage Age
#> 1 1.815365       20          30  60