Nitrogen fixing by alder plants

These data were collected by biologist Mike Anderson in a study of nitrogen fixation by bacteria growing on the root nodules of alder bushes.

data(Alder)

Format

A data frame Alder with 196 rows and 24 variables:

LAND. landscape, floodplain vs. upland.

SAMPPER. sampling period: early, mid, late

SPECIES. host species, Alnus tenuifolia (AT) vs. A. crispa (AC)

STAGE. successional stage, early vs. late in floodplain and upland landscapes. This is not equivalent across landscapes.

JULDAY. Julian day

PERNODN. nodule percent nitrogen by mass

RF. bacterial genotype

SNF. nitrogen fixation rate of nodule tissue, umol N2/gram of nodule dry weight/hr

SLA. specific leaf weight, grams of leaf weight/square-meter, dry

ONECM. soil temperature at 1 cm depth

FIVECM. soil temperature at 5 cm depth

PERH2O. soil moisture, percent H2O by mass

DEL. del15N of leaf tissue

DELNOD. del15N of nodule tissue

NPERAREA. leaf nitrogen content per unit leaf area

NDiff. nitrogen content difference between leaf and nodule of the same plant

delDiff. del15N difference between leaf and nodule of the same plant

SITE. Site designations: 1A,B,C for replicate early succession floodplain sites, 4A,B,C for late succession floodplain, UP1A,B,C for early succession upland and UP3A,B,C for late succession upland

HABSPEC. habitat+species, concatenated LAND, STAGE, SPECIES

SITESPEC. concatenated SITE, SPECIES

REP. replicate site within a given level of HABSPEC

PLNO. plant number, unique for individuals of each species (AT1-180, AC1-270)

Source

Michael Anderson

Details

Two questions that Anderson wanted to answer are: (1) Can any variation in nitrogen fixation (variable SNF) be attributed to genotype (variable RF)? (2) What are the major sources of variation in SNF and PERLEAFN? Variables of biological interest are seasonality (SAMPPER or JULDAY), soil temperature and moisture, and habitat differences (STAGE for host species AT andSTAGE and LAND for host species AC).

Three replicate sites were sampled for each landscape/stage combination in three sampling periods across the growing season. Site sampling was arranged in a Latin Square design in order to systematize any effects of seasonality on N2-fixation rates.

References

Anderson MD, Ruess RW, Myrold DD, Taylor DL. ``Host species and habitat affect modulation by specific Frankia genotypes in interior Alaska'' Oecologia (2009) 160:619-630.

Examples

mod <- lm(logSNF ~ RF + SITESPEC, data = Alder)