Create bootstrapped ensembles of a model

Create bootstrapped ensembles of a model

mod_ensemble(model, nreps = 2, data = NULL)

Arguments

model
a model whose data will be used for resampling

nreps
how many resampling trials should be created

data
a data table to use for resampling. This is not needed for many common model types, such as lm, glm, etc. See details.

Details

The approach to bootstrapping implemented by this function is to create a set of bootstrap trials all in one go. Then, other functions such as mod_effect() and mod_eval() will be used to extract the information from each of the bootstrap replicates. Many model types in R carry the data used to train the model as part of the model object produced. For these types of models, e.g. lm and glm, there is no need to provide a value for the data argument. But there are some types of models for which the training data cannot be extracted from the model object. In such situations, you use data = to provide the data set to use for resampling.